% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.character.matrix.r
\name{align.table}
\alias{align.table}
\title{align.table}
\usage{
align.table(x, align, space = 1)
}
\arguments{
\item{x}{x}

\item{align}{align}

\item{space}{space}
}
\description{
align.table
}
\keyword{internal}
