% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.r
\name{as.list.matrix}
\alias{as.list.matrix}
\title{as.list.matrix}
\usage{
\method{as.list}{matrix}(x, byrow = TRUE, ...)
}
\arguments{
\item{x}{x}

\item{byrow}{byrow}

\item{...}{...}
}
\description{
as.list.matrix
}
\keyword{internal}
