\name{ascii.CrossTable}
\alias{ascii.CrossTable}
\title{ascii methodd for class CrossTable...}
\usage{\method{ascii}{CrossTable} (x, ...)}
\description{ascii methodd for class CrossTable}
\details{\code{descr} package}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{A crosstable object}
\item{...}{Additional arguments.  (see \code{?ascii.default}).}}
