\name{ascii.describe}
\alias{ascii.describe}
\title{ascii method fo class describe...}
\usage{\method{ascii}{describe}(x, condense=TRUE, ...)
}
\description{ascii method fo class describe}
\details{\code{Hmisc} package}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{An R object of class found among \code{methods(ascii)}.}
\item{condense}{default is TRUE to condense the output with regard to the 5
lowest and highest values and the frequency table (\code{describe()} in
package \code{Hmisc}).}
\item{...}{Additional arguments.  (see \code{?ascii.default}).}
}

