% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiDescr.r
\name{ascii.microbenchmark}
\alias{ascii.microbenchmark}
\title{Ascii formatting for a microbenchmark}
\usage{
\method{ascii}{microbenchmark}(x, unit, order, signif, row.names = FALSE, caption = NULL, ...)
}
\arguments{
\item{x}{an object of class 'microbenchmark'}

\item{unit}{What unit to print the timings in. Default value taken
from the option 'microbenchmark.unit'}

\item{order}{If present, order results according to this column of
the output.}

\item{signif}{If present, limit the limit of significant digits
shown.}

\item{row.names}{Argument passed to ascii}

\item{caption}{logical; if not NULL, then add caption with units
specified; otherwise, add units as part of an asciiMixed
object.}

\item{...}{Other parameters to pass to ascii for the summary table}
}
\value{
ascii object
}
\description{
The default implementation returns an asciiMixed object with the
units for the first element.
}
