\name{ascii.sessionInfo}
\alias{ascii.sessionInfo}
\title{ascii metho for class sessionInfo...}
\usage{\method{ascii}{sessionInfo}(x, locale=TRUE, ...)
}
\description{ascii metho for class sessionInfo}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{An R object of class found among \code{methods(ascii)}.}
\item{locale}{default is TRUE to show locale information
(\code{sessionInfo()}).}
\item{...}{Additional arguments.  (Currently ignored.)}
}

