\name{ascii.stat.table}
\alias{ascii.stat.table}
\title{ascii method for class stat...}
\usage{\method{ascii}{stat.table}(x, tgroup, lgroup, ...)
}
\description{ascii method for class stat.table}
\details{\code{Epi} package}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{An R object of class \code{stat.table}.}
\item{tgroup}{Character vector or a list of character vectors defining
major top column headings.  The default is to have none (\code{NULL}).}
\item{lgroup}{Character vector or list of character vectors defining major
left row headings.  The default is to have none (\code{NULL}).}
\item{...}{Additional arguments.  (see \code{?ascii.default}.)}
}

