\name{ascii.zoo}
\alias{ascii.zoo}
\title{ascii method for class zoo...}
\usage{\method{ascii}{zoo} (x, ...)}
\description{ascii method for class zoo}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{An R object of class found among \code{methods(ascii)}.}
\item{...}{Additional arguments.  (see \code{?ascii.default}).}}
