\name{Asciidoc}
\alias{Asciidoc}
\title{Sweave wrapper for asciidoc...}
\description{Sweave wrapper for asciidoc
Like \code{Sweave}, but use \code{RweaveAsciidoc} driver as default value.}
\author{David Hajage \email{dhajage@gmail.com}}
\seealso{\code{\link{Sweave}}}
\keyword{IO}
\keyword{file}
\arguments{\item{file}{Name of Sweave source file.}
\item{driver}{Sweave driver}
\item{syntax}{Sweave syntax}
\item{...}{Further arguments passed to the driver's setup function.}}
