% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\docType{class}
\name{asciiList-class}
\alias{asciiList-class}
\alias{asciiList}
\title{ascii list generator}
\description{
ascii list generator
}
\section{Methods}{

\describe{
\item{\code{show.asciidoc(
  x = .self$x,
  caption = .self$caption,
  caption.level = .self$caption.level,
  list.type = .self$list.type
)}}{print a list with asciidoc markup}

\item{\code{show.org(
  x = .self$x,
  caption = .self$caption,
  caption.level = .self$caption.level,
  list.type = .self$list.type
)}}{print a list with org markup}

\item{\code{show.pandoc(
  x = .self$x,
  caption = .self$caption,
  caption.level = .self$caption.level,
  list.type = .self$list.type
)}}{print a list with pandoc markup}

\item{\code{show.rest(
  x = .self$x,
  caption = .self$caption,
  caption.level = .self$caption.level,
  list.type = .self$list.type
)}}{print a list with rest markup}

\item{\code{show.t2t(
  x = .self$x,
  caption = .self$caption,
  caption.level = .self$caption.level,
  list.type = .self$list.type
)}}{print a list with t2t markup}

\item{\code{show.textile(
  x = .self$x,
  caption = .self$caption,
  caption.level = .self$caption.level,
  list.type = .self$list.type
)}}{print a list with textile markup}
}}

\author{
David Hajage
}
