% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\docType{class}
\name{asciiMixed-class}
\alias{asciiMixed-class}
\alias{asciiMixed}
\title{ascii mixed generator}
\description{
ascii mixed generator
}
\section{Methods}{

\describe{
\item{\code{show.asciidoc()}}{print everything with asciidoc markup}

\item{\code{show.org()}}{print everything with org markup}

\item{\code{show.pandoc()}}{print everything with pandoc markup}

\item{\code{show.rest()}}{print everything with rest markup}

\item{\code{show.t2t()}}{print everything with t2t markup}

\item{\code{show.textile()}}{print everything with textile markup}
}}

\author{
David Hajage
}
