\name{asciiOpts}
\alias{asciiOpts}
\title{asciiOptions}
\usage{
asciiOpts(select = "all", .backends = NULL, .outputs = NULL, .extensions = NULL, .cli = NULL, .args = NULL, .O = NULL, .f = NULL, .e = NULL, .preambule = NULL)
}
\arguments{
  \item{select}{select}

  \item{.backends}{.backends}

  \item{.outputs}{.outputs}

  \item{.extensions}{.extensions}

  \item{.cli}{.cli}

  \item{.args}{.args}

  \item{.O}{.O}

  \item{.f}{.f}

  \item{.e}{.e}

  \item{.preambule}{.preambule}
}
\value{
options
}
\description{
asciiOptions
}
\author{
David.hajage
}
\keyword{internal}

