% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\docType{class}
\name{asciiTable-class}
\alias{asciiTable-class}
\alias{asciiTable}
\title{ascii table generator}
\description{
ascii table generator
}
\section{Methods}{

\describe{
\item{\code{show.asciidoc(
  x = .self$x,
  include.rownames = .self$include.rownames,
  include.colnames = .self$include.colnames,
  rownames = .self$rownames,
  colnames = .self$colnames,
  format = .self$format,
  digits = .self$digits,
  decimal.mark = .self$decimal.mark,
  na.print = .self$na.print,
  caption = .self$caption,
  caption.level = .self$caption.level,
  width = .self$width,
  frame = .self$frame,
  grid = .self$grid,
  valign = .self$valign,
  header = .self$header,
  footer = .self$footer,
  align = .self$align,
  col.width = .self$col.width,
  style = .self$style,
  lgroup = .self$lgroup,
  n.lgroup = .self$n.lgroup,
  lalign = .self$lalign,
  lvalign = .self$lvalign,
  lstyle = .self$lstyle,
  rgroup = .self$rgroup,
  n.rgroup = .self$n.rgroup,
  ralign = .self$ralign,
  rvalign = .self$rvalign,
  rstyle = .self$rstyle,
  tgroup = .self$tgroup,
  n.tgroup = .self$n.tgroup,
  talign = .self$talign,
  tvalign = .self$tvalign,
  tstyle = .self$tstyle,
  bgroup = .self$bgroup,
  n.bgroup = .self$n.bgroup,
  balign = .self$balign,
  bvalign = .self$bvalign,
  bstyle = .self$bstyle
)}}{print a table with asciidoc markup}

\item{\code{show.org(
  x = .self$x,
  include.rownames = .self$include.rownames,
  include.colnames = .self$include.colnames,
  rownames = .self$rownames,
  colnames = .self$colnames,
  format = .self$format,
  digits = .self$digits,
  decimal.mark = .self$decimal.mark,
  na.print = .self$na.print,
  caption = .self$caption,
  caption.level = .self$caption.level,
  width = .self$width,
  frame = .self$frame,
  grid = .self$grid,
  valign = .self$valign,
  header = .self$header,
  footer = .self$footer,
  align = .self$align,
  col.width = .self$col.width,
  style = .self$style,
  lgroup = .self$lgroup,
  n.lgroup = .self$n.lgroup,
  lalign = .self$lalign,
  lvalign = .self$lvalign,
  lstyle = .self$lstyle,
  rgroup = .self$rgroup,
  n.rgroup = .self$n.rgroup,
  ralign = .self$ralign,
  rvalign = .self$rvalign,
  rstyle = .self$rstyle,
  tgroup = .self$tgroup,
  n.tgroup = .self$n.tgroup,
  talign = .self$talign,
  tvalign = .self$tvalign,
  tstyle = .self$tstyle,
  bgroup = .self$bgroup,
  n.bgroup = .self$n.bgroup,
  balign = .self$balign,
  bvalign = .self$bvalign,
  bstyle = .self$bstyle
)}}{print a table with org-mode markup}

\item{\code{show.pandoc(
  x = .self$x,
  include.rownames = .self$include.rownames,
  include.colnames = .self$include.colnames,
  rownames = .self$rownames,
  colnames = .self$colnames,
  format = .self$format,
  digits = .self$digits,
  decimal.mark = .self$decimal.mark,
  na.print = .self$na.print,
  caption = .self$caption,
  caption.level = .self$caption.level,
  width = .self$width,
  frame = .self$frame,
  grid = .self$grid,
  valign = .self$valign,
  header = .self$header,
  footer = .self$footer,
  align = .self$align,
  col.width = .self$col.width,
  style = .self$style,
  lgroup = .self$lgroup,
  n.lgroup = .self$n.lgroup,
  lalign = .self$lalign,
  lvalign = .self$lvalign,
  lstyle = .self$lstyle,
  rgroup = .self$rgroup,
  n.rgroup = .self$n.rgroup,
  ralign = .self$ralign,
  rvalign = .self$rvalign,
  rstyle = .self$rstyle,
  tgroup = .self$tgroup,
  n.tgroup = .self$n.tgroup,
  talign = .self$talign,
  tvalign = .self$tvalign,
  tstyle = .self$tstyle,
  bgroup = .self$bgroup,
  n.bgroup = .self$n.bgroup,
  balign = .self$balign,
  bvalign = .self$bvalign,
  bstyle = .self$bstyle
)}}{print a table with pandoc markup}

\item{\code{show.rest(
  x = .self$x,
  include.rownames = .self$include.rownames,
  include.colnames = .self$include.colnames,
  rownames = .self$rownames,
  colnames = .self$colnames,
  format = .self$format,
  digits = .self$digits,
  decimal.mark = .self$decimal.mark,
  na.print = .self$na.print,
  caption = .self$caption,
  caption.level = .self$caption.level,
  width = .self$width,
  frame = .self$frame,
  grid = .self$grid,
  valign = .self$valign,
  header = .self$header,
  footer = .self$footer,
  align = .self$align,
  col.width = .self$col.width,
  style = .self$style,
  lgroup = .self$lgroup,
  n.lgroup = .self$n.lgroup,
  lalign = .self$lalign,
  lvalign = .self$lvalign,
  lstyle = .self$lstyle,
  rgroup = .self$rgroup,
  n.rgroup = .self$n.rgroup,
  ralign = .self$ralign,
  rvalign = .self$rvalign,
  rstyle = .self$rstyle,
  tgroup = .self$tgroup,
  n.tgroup = .self$n.tgroup,
  talign = .self$talign,
  tvalign = .self$tvalign,
  tstyle = .self$tstyle,
  bgroup = .self$bgroup,
  n.bgroup = .self$n.bgroup,
  balign = .self$balign,
  bvalign = .self$bvalign,
  bstyle = .self$bstyle
)}}{print a table with restructuredText markup}

\item{\code{show.t2t(
  x = .self$x,
  include.rownames = .self$include.rownames,
  include.colnames = .self$include.colnames,
  rownames = .self$rownames,
  colnames = .self$colnames,
  format = .self$format,
  digits = .self$digits,
  decimal.mark = .self$decimal.mark,
  na.print = .self$na.print,
  caption = .self$caption,
  caption.level = .self$caption.level,
  width = .self$width,
  frame = .self$frame,
  grid = .self$grid,
  valign = .self$valign,
  header = .self$header,
  footer = .self$footer,
  align = .self$align,
  col.width = .self$col.width,
  style = .self$style,
  lgroup = .self$lgroup,
  n.lgroup = .self$n.lgroup,
  lalign = .self$lalign,
  lvalign = .self$lvalign,
  lstyle = .self$lstyle,
  rgroup = .self$rgroup,
  n.rgroup = .self$n.rgroup,
  ralign = .self$ralign,
  rvalign = .self$rvalign,
  rstyle = .self$rstyle,
  tgroup = .self$tgroup,
  n.tgroup = .self$n.tgroup,
  talign = .self$talign,
  tvalign = .self$tvalign,
  tstyle = .self$tstyle,
  bgroup = .self$bgroup,
  n.bgroup = .self$n.bgroup,
  balign = .self$balign,
  bvalign = .self$bvalign,
  bstyle = .self$bstyle
)}}{print a table with txt2tags markup}

\item{\code{show.textile(
  x = .self$x,
  include.rownames = .self$include.rownames,
  include.colnames = .self$include.colnames,
  rownames = .self$rownames,
  colnames = .self$colnames,
  format = .self$format,
  digits = .self$digits,
  decimal.mark = .self$decimal.mark,
  na.print = .self$na.print,
  caption = .self$caption,
  caption.level = .self$caption.level,
  width = .self$width,
  frame = .self$frame,
  grid = .self$grid,
  valign = .self$valign,
  header = .self$header,
  footer = .self$footer,
  align = .self$align,
  col.width = .self$col.width,
  style = .self$style,
  lgroup = .self$lgroup,
  n.lgroup = .self$n.lgroup,
  lalign = .self$lalign,
  lvalign = .self$lvalign,
  lstyle = .self$lstyle,
  rgroup = .self$rgroup,
  n.rgroup = .self$n.rgroup,
  ralign = .self$ralign,
  rvalign = .self$rvalign,
  rstyle = .self$rstyle,
  tgroup = .self$tgroup,
  n.tgroup = .self$n.tgroup,
  talign = .self$talign,
  tvalign = .self$tvalign,
  tstyle = .self$tstyle,
  bgroup = .self$bgroup,
  n.bgroup = .self$n.bgroup,
  balign = .self$balign,
  bvalign = .self$bvalign,
  bstyle = .self$bstyle
)}}{print a table with textile markup}
}}

\author{
David Hajage
}
