% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.asciidoc.r
\name{beauty.asciidoc}
\alias{beauty.asciidoc}
\title{beauty.asciidoc}
\usage{
beauty.asciidoc(x, beauti = c("e", "m", "s"))
}
\arguments{
\item{x}{x}

\item{beauti}{beauti}
}
\description{
beauty.asciidoc
}
\keyword{internal}
