% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.pandoc.r
\name{beauty.pandoc}
\alias{beauty.pandoc}
\title{beauty.pandoc}
\usage{
beauty.pandoc(x, beauti = c("e", "m", "s"))
}
\arguments{
\item{x}{x}

\item{beauti}{beauti}
}
\description{
beauty.pandoc
}
\keyword{internal}
