% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.textile.r
\name{beauty.textile}
\alias{beauty.textile}
\title{beauty.textile}
\usage{
beauty.textile(x, beauti = c("e", "m", "s"))
}
\arguments{
\item{x}{x}

\item{beauti}{beauti}
}
\description{
beauty.textile
}
\keyword{internal}
