\name{cacheSweaveAsciiSetup}
\alias{cacheSweaveAsciiSetup}
\title{A driver to parse ascii noweb files with Sweave tool - cacheSweave based}
\usage{
cacheSweaveAsciiSetup(..., cache = FALSE, trace = FALSE, dependson = NULL)
}
\description{
A driver to parse ascii noweb files with Sweave tool - cacheSweave based
}
\author{
David Hajage
}
\keyword{internal}

