\name{export}
\alias{export}
\alias{convert}
\alias{section}
\alias{print.section}
\alias{sexpr}
\alias{print.sexpr}
\alias{paragraph}
\alias{print.paragraph}

\title{Use asciidoc from R}
\description{
  These functions allow to use asciidoc and a2x toolchains directly from R.
}
\usage{
  export(..., file = NULL, format = "xhtml", open = NULL, main = NULL,
         author = NULL, email = NULL, revdate = NULL, revnumber = NULL,
         encoding = NULL, cmd = NULL, cygwin = TRUE)
 }
\arguments{
  \item{...}{\code{ascii}, \code{section}, \code{paragraph}, \code{verbatim} or
    \code{sexpr} object.}
  
}
\details{
}
\value{
}
\author{David Hajage \email{dhajage@gmail.com}}
\examples{
}
\keyword{print}
