% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocharac.r
\name{formatCustom}
\alias{formatCustom}
\title{formatCustom}
\usage{
formatCustom(x, format = "nice", digits = 2, ...)
}
\arguments{
\item{x}{x}

\item{format}{format}

\item{digits}{digits}

\item{...}{...}
}
\description{
formatCustom
}
\keyword{internal}
