\name{graph}
\alias{graph}
\title{Insert graph...}
\usage{graph(graph)
}
\description{Insert graph}
\details{\code{graph} can be used with \code{export} function to insert an R graphic.}
\value{An out object}
\author{David Hajage}
\arguments{\item{graph}{character string (a link to a graphic file)}
}

