\name{header.asciidoc}
\alias{header.asciidoc}
\title{header...}
\usage{header.asciidoc(caption=NULL, caption.level=NULL, frame=NULL,
    grid=NULL, col.width=1, width=0)}
\description{header.asciidoc}
\keyword{internal}
\arguments{\item{caption}{caption}
\item{caption.level}{caption.level}
\item{frame}{frame}
\item{grid}{grid}
\item{col.width}{col.width}
\item{width}{width}}
