% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.pandoc.r
\name{header.pandoc}
\alias{header.pandoc}
\title{header.pandoc}
\usage{
header.pandoc(caption = NULL, caption.level = NULL)
}
\arguments{
\item{caption}{caption}

\item{caption.level}{caption.level}
}
\description{
header.pandoc
}
\keyword{internal}
