% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.rest.r
\name{header.rest}
\alias{header.rest}
\title{header.rest}
\usage{
header.rest(caption = NULL, caption.level = NULL)
}
\arguments{
\item{caption}{caption}

\item{caption.level}{caption.level}
}
\description{
header.rest
}
\keyword{internal}
