% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.t2t.r
\name{header.t2t}
\alias{header.t2t}
\title{header.t2t}
\usage{
header.t2t(caption = NULL, caption.level = "")
}
\arguments{
\item{caption}{caption}

\item{caption.level}{caption.level}
}
\description{
header.t2t
}
\keyword{internal}
