\name{header.textile}
\alias{header.textile}
\title{header.textile}
\usage{
header.textile(caption = NULL, caption.level = NULL, frame = NULL, width = 0)
}
\arguments{
  \item{caption}{caption}

  \item{caption.level}{caption.level}

  \item{frame}{frame}

  \item{width}{width}
}
\description{
header.textile
}
\keyword{internal}

