% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.r
\name{interleave.matrix}
\alias{interleave.matrix}
\title{interleave.matrix}
\usage{
interleave.matrix(..., byrow = TRUE)
}
\arguments{
\item{...}{...}

\item{byrow}{byrow}
}
\description{
interleave.matrix
}
\keyword{internal}
