\name{makeCacheSweaveAsciiCodeRunner}
\alias{makeCacheSweaveAsciiCodeRunner}
\title{A driver to parse ascii noweb files with Sweave tool - cacheSweave based}
\usage{
makeCacheSweaveAsciiCodeRunner(evalFunc = cacheSweave:::cacheSweaveEvalWithOpt)
}
\description{
A driver to parse ascii noweb files with Sweave tool - cacheSweave based
}
\author{
David Hajage
}
\keyword{internal}

