% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RweaveAscii.r
\name{makeRweaveAsciiCodeRunner}
\alias{makeRweaveAsciiCodeRunner}
\title{makeRweaveAsciiCodeRunner}
\usage{
makeRweaveAsciiCodeRunner(evalFunc = RweaveEvalWithOpt)
}
\arguments{
\item{evalFunc}{evalFunc}
}
\description{
makeRweaveAsciiCodeRunner
}
\keyword{internal}
