\name{math}
\alias{math}
\title{Insert an equation...}
\usage{math(x, notation="latexmath")}
\description{Insert an equation}
\details{\code{math} can be used with \code{export} function to insert some maths}
\author{David Hajage}
\arguments{\item{x}{some maths}
\item{notation}{latexmath or asciimath}}
