\name{out}
\alias{out}
\title{Export R objects}
\usage{
out(x, results = "verbatim")
}
\arguments{
  \item{x}{an R object}

  \item{results}{if \code{'verbatim'}, the output is
  included in a verbatim environment. If \code{'ascii'},
  the output is taken to be already proper markup and
  included as is.}
}
\value{
An out object
}
\description{
\code{out} can be used with \code{export} function to insert an R results
}
\author{
David Hajage
}

