\name{Pandoc}
\alias{Pandoc}
\title{Sweave wrapper for pandoc...}
\description{Sweave wrapper for pandoc
Like \code{Sweave}, but use \code{RweavePandoc} driver as default value.}
\author{David Hajage \email{dhajage@gmail.com} Matti Pastell \email{matti.pastell@helsinki.fi}}
\seealso{\code{\link{Sweave}}}
\keyword{IO}
\keyword{file}
\arguments{\item{file}{Name of Sweave source file.}
\item{driver}{Sweave driver}
\item{syntax}{Sweave syntax}
\item{...}{Further arguments passed to the driver's setup function.}
}

