\name{paragraph}
\alias{paragraph}
\title{Create a paragraph}
\usage{
paragraph(..., new = TRUE)
}
\arguments{
  \item{...}{strings composing the paragraph}

  \item{new}{whether to create a new paragraph or to
  continue a preceding one}
}
\value{
A paragraph object.
}
\description{
\code{paragraph} can be used with \code{export} function to add\dots a paragraph
}
\author{
David Hajage
}

