\name{paste.matrix}
\alias{paste.matrix}
\title{paste.matrix}
\usage{
paste.matrix(..., sep = " ", transpose.vector = FALSE, collapse = NULL, byrow = FALSE)
}
\arguments{
  \item{...}{...}

  \item{sep}{sep}

  \item{transpose.vector}{transpose.vector}

  \item{collapse}{collapse}

  \item{byrow}{byrow}
}
\description{
paste.matrix
}
\keyword{internal}

