\name{print.character.matrix}
\alias{print.character.matrix}
\title{print...}
\usage{print.character.matrix(x, vsep="|", before_vsep="", after_vsep="",
    hsep="-", csep="+", before_cell_content=" ", after_cell_content=" ",
    line_separator=TRUE, line_separator_pos=NULL, justify=c("left",
    "right", "centre", "none"), right_alignment=FALSE, print=TRUE)}
\description{print.character.matrix}
\keyword{internal}
\arguments{\item{x}{x}
\item{vsep}{vsep}
\item{before_vsep}{before_vsep}
\item{after_vsep}{after_vsep}
\item{hsep}{hsep}
\item{csep}{csep}
\item{before_cell_content}{before_cell_content}
\item{after_cell_content}{after_cell_content}
\item{line_separator}{line_separator}
\item{line_separator_pos}{line_separator_pos}
\item{justify}{justify}
\item{right_alignment}{right_alignment}
\item{print}{print}}
