\name{print.fig}
\alias{print.fig}
\title{Print an graph object}
\usage{
\method{print}{fig}(x, backend = getOption("asciiBackend"), ...)
}
\arguments{
  \item{x}{an graph object}

  \item{backend}{ascii backend}

  \item{...}{not used}
}
\description{
Print an graph object
}
\author{
David Hajage
}

