\name{print.out}
\alias{print.out}
\title{Print an out object...}
\usage{print.out(x, backend=getOption("asciiBackend"), ...)
}
\description{Print an out object}
\details{Print an out object}
\author{David Hajage}
\arguments{\item{x}{an out object}
\item{backend}{ascii backend}
\item{...}{not used}
}

