% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{print.paragraph}
\alias{print.paragraph}
\title{Print a paragraph object}
\usage{
\method{print}{paragraph}(x, ...)
}
\arguments{
\item{x}{a paragraph object}

\item{...}{not used}
}
\description{
Print a paragraph object
}
\author{
David Hajage
}
