\name{print.Report}
\alias{print.Report}
\title{Print method for class Report...}
\usage{print.Report(x, help=FALSE, ...)
}
\description{Print method for class \code{Report}}
\details{Display report characteristics}
\author{David Hajage}
\arguments{\item{x}{a report (generated with \code{Report$new()})}
\item{help}{logical print help?}
\item{...}{Not used}
}

