\name{print.ReportClass}
\alias{print.ReportClass}
\title{Print method for class ReportClass...}
\usage{print.ReportClass(x, ...)
}
\description{Print method for class \code{ReportClass}}
\details{Display Report help}
\author{David Hajage}
\arguments{\item{x}{the ReportClass}
\item{...}{Not used}
}

