\name{print.section}
\alias{print.section}
\title{Print a section object...}
\usage{print.section(x, type=getOption("asciiType"), ...)
}
\description{Print a section object}
\details{Print a section object}
\author{David Hajage}
\arguments{\item{x}{a section object}
\item{type}{ascii type}
\item{...}{not used}
}

