% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{print.sexpr}
\alias{print.sexpr}
\title{Print a sexpr object}
\usage{
\method{print}{sexpr}(x, ...)
}
\arguments{
\item{x}{a sexpr object}

\item{...}{not used}
}
\description{
Print a sexpr object
}
\author{
David Hajage
}
