\name{rbind.ascii}
\alias{rbind.ascii}
\title{Rbind two ascii objects}
\usage{rbind.ascii(x, y, caption=NULL, caption.level=NULL, frame=NULL,
    grid=NULL, col.width=1, width=0)}
\description{Rbind two ascii objects}
\details{This function binds rows of two ascii table.}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{an ascii table}
\item{y}{another ascii table}
\item{caption}{see \code{?ascii}}
\item{caption.level}{see \code{?ascii}}
\item{frame}{see \code{?ascii}}
\item{grid}{see \code{?ascii}}
\item{col.width}{see \code{?ascii}}
\item{width}{see \code{?ascii}}}
