% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RweaveAscii.r
\name{RtangleAsciiWritedoc}
\alias{RtangleAsciiWritedoc}
\title{RtangleAsciiWritedoc}
\usage{
RtangleAsciiWritedoc(object, chunk)
}
\arguments{
\item{object}{object}

\item{chunk}{chunk}
}
\description{
RtangleAsciiWritedoc
}
\keyword{internal}
