% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RweaveAscii.r
\name{RweaveAsciiOptions}
\alias{RweaveAsciiOptions}
\title{RweaveAsciiOptions}
\usage{
RweaveAsciiOptions(options)
}
\arguments{
\item{options}{options}
}
\description{
RweaveAsciiOptions
}
\keyword{internal}
