\name{RweaveAsciiSetup}
\alias{RweaveAsciiSetup}
\title{RweaveAsciiSetup}
\usage{
RweaveAsciiSetup(file, syntax, output = NULL, quiet = FALSE, debug = FALSE, extension = "txt", backend = "asciidoc", openSchunk = "----", closeSchunk = "\n----\n", openSinput = "", closeSinput = "", openSoutput = "\n", closeSoutput = "", indent = "", openInclude = "include::", closeInclude = ".txt[]", openFig = "image::", closeFig = "[]", ...)
}
\arguments{
  \item{file}{file}

  \item{syntax}{syntax}

  \item{output}{output}

  \item{quiet}{quite}

  \item{debug}{debug}

  \item{stylepath}{stylepath}

  \item{...}{...}
}
\description{
RweaveAsciiSetup
}
\keyword{internal}

