% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RweaveAscii.r
\name{RweaveAsciiWritedoc}
\alias{RweaveAsciiWritedoc}
\title{RweaveAsciiWritedoc}
\usage{
RweaveAsciiWritedoc(object, chunk)
}
\arguments{
\item{object}{object}

\item{chunk}{chunk}
}
\description{
RweaveAsciiWritedoc
}
\keyword{internal}
