\name{section}
\alias{section}
\title{Create a section}
\usage{
section(caption, caption.level = 1)
}
\arguments{
  \item{caption}{a string}

  \item{caption.level}{caption level}
}
\value{
A section object.
}
\description{
\code{section} can be used with \code{export} function to add\dots a section
}
\author{
David Hajage
}

