\name{sexpr}
\alias{sexpr}
\title{Insert an inline R result...}
\usage{sexpr(x)}
\description{Insert an inline R result}
\details{\code{sexpr} can be used with \code{export} function to insert an inline R results}
\value{A sexpr object.}
\author{David Hajage}
\arguments{\item{x}{an R results (of length one)}}
