\name{show.asciidoc.table}
\alias{show.asciidoc.table}
\title{show...}
\usage{show.asciidoc.table(x, include.rownames=FALSE, include.colnames=FALSE, rownames, colnames,
    format="f", digits=2, decimal.mark=".", na.print="", caption,
    caption.level, width=0, frame, grid, valign, header=FALSE,
    footer=FALSE, align, col.width=1, style, lgroup, n.lgroup,
    lalign="c", lvalign="middle", lstyle="h", rgroup, n.rgroup,
    ralign="c", rvalign="middle", rstyle="h", tgroup, n.tgroup,
    talign="c", tvalign="middle", tstyle="h", bgroup, n.bgroup,
    balign="c", bvalign="middle", bstyle="h", ...)
}
\description{show.asciidoc.table}
\keyword{internal}
\arguments{\item{x}{x}
\item{include.rownames}{include.rownames}
\item{include.colnames}{include.colnames}
\item{rownames}{rownames}
\item{colnames}{colnames}
\item{format}{format}
\item{digits}{digits}
\item{decimal.mark}{decimal.mark}
\item{na.print}{na.print}
\item{caption}{caption}
\item{caption.level}{}
\item{width}{width}
\item{frame}{frame}
\item{grid}{grid}
\item{valign}{valign}
\item{header}{header}
\item{footer}{footer}
\item{align}{align}
\item{col.width}{col.width}
\item{style}{style}
\item{lgroup}{lgroup}
\item{n.lgroup}{n.lgroup}
\item{lalign}{lalign}
\item{lvalign}{lvalign}
\item{lstyle}{lstyle}
\item{rgroup}{rgroup}
\item{n.rgroup}{n.rgroup}
\item{ralign}{ralign}
\item{rvalign}{rvalign}
\item{rstyle}{rstyle}
\item{tgroup}{tgroup}
\item{n.tgroup}{n.tgroup}
\item{talign}{talign}
\item{tvalign}{tvalign}
\item{tstyle}{tstyle}
\item{bgroup}{bgroup}
\item{n.bgroup}{n.bgroup}
\item{balign}{balign}
\item{bvalign}{bvalign}
\item{bstyle}{bstyle}
\item{...}{...}
}

