\name{show.t2t.list}
\alias{show.t2t.list}
\title{show...}
\usage{show.t2t.list(x, caption=NULL, caption.level=NULL, list.type="bullet",
    ...)}
\description{show.t2t.list}
\keyword{internal}
\arguments{\item{x}{x}
\item{caption}{caption}
\item{caption.level}{caption.level}
\item{list.type}{list.type}
\item{...}{...}}
