\name{show.textile.list}
\alias{show.textile.list}
\title{show.textile.list}
\usage{
show.textile.list(x, caption = NULL, caption.level = NULL, list.type = "bullet", ...)
}
\arguments{
  \item{x}{x}

  \item{caption}{caption}

  \item{caption.level}{caption.level}

  \item{list.type}{list.type}

  \item{...}{...}
}
\description{
show.textile.list
}
\keyword{internal}

