\name{tocharac}
\alias{tocharac}
\title{tocharac}
\usage{
tocharac(x, include.rownames = FALSE, include.colnames = FALSE, rownames = NULL, colnames = NULL, format = "f", digits = 2, decimal.mark = ".", na.print = "")
}
\arguments{
  \item{x}{x}

  \item{include.rownames}{include.rownames}

  \item{include.colnames}{include.colnames}

  \item{rownames}{rownames}

  \item{colnames}{colnames}

  \item{format}{format}

  \item{digits}{digits}

  \item{decimal.mark}{decimal.mark}

  \item{na.print}{na.print}
}
\description{
tocharac
}
\keyword{internal}

