\name{verbatim}
\alias{verbatim}
\title{Create a verbatim paragraph}
\usage{
verbatim(...)
}
\arguments{
  \item{...}{strings composing the paragraph (line by
  line)}
}
\value{
A verbatim object.
}
\description{
\code{verbatim} can be used with \code{export} function to add a verbatim paragraph
}
\author{
David Hajage
}

