% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.asciidoc.r
\name{vsep.asciidoc}
\alias{vsep.asciidoc}
\title{vsep.asciidoc}
\usage{
vsep.asciidoc(align = NULL, valign = NULL, style = NULL)
}
\arguments{
\item{align}{align}

\item{valign}{valign}

\item{style}{style}
}
\description{
vsep.asciidoc
}
\keyword{internal}
