% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.textile.r
\name{vsep.textile}
\alias{vsep.textile}
\title{vsep.textile}
\usage{
vsep.textile(align = NULL, valign = NULL, style = NULL)
}
\arguments{
\item{align}{align}

\item{valign}{valign}

\item{style}{style}
}
\description{
vsep.textile
}
\keyword{internal}
