% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverAscii.r
\name{weaverAsciiFinish}
\alias{weaverAsciiFinish}
\title{weaverAsciiFinish}
\usage{
weaverAsciiFinish(object, error = FALSE)
}
\arguments{
\item{object}{object}

\item{error}{error}
}
\description{
weaverAsciiFinish
}
\keyword{internal}
