\name{asciiOpts}
\alias{asciiOpts}
\title{asciiOptions...}
\usage{asciiOpts(select="all", .backends, .outputs, .extensions, .cli, .args, .O, .f,
    .e, .preambule)
}
\description{asciiOptions}
\value{options}
\author{David.hajage}
\keyword{internal}
\arguments{\item{select}{select}
\item{.backends}{.backends}
\item{.outputs}{.outputs}
\item{.extensions}{.extensions}
\item{.cli}{.cli}
\item{.args}{.args}
\item{.O}{.O}
\item{.f}{.f}
\item{.e}{.e}
\item{.preambule}{.preambule}
}

