\name{cbind.ascii}
\alias{cbind.ascii}
\title{Cbind two ascii objects}
\usage{cbind.ascii(..., caption=NULL, caption.level=NULL, frame=NULL, grid=NULL,
    col.width=1, width=0)
}
\description{Cbind two ascii objects}
\details{This function binds cols of two ascii table.}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{...}{ascii objects}
\item{caption}{see \code{?ascii}}
\item{caption.level}{see \code{?ascii}}
\item{frame}{see \code{?ascii}}
\item{grid}{see \code{?ascii}}
\item{col.width}{see \code{?ascii}}
\item{width}{see \code{?ascii}}
}

