\name{cbind.ascii}
\alias{cbind.ascii}
\title{Cbind two ascii objects}
\usage{cbind.ascii(x, y, caption, caption.level, frame, grid, col.width=1,
    width=0)}
\description{Cbind two ascii objects}
\details{This function binds cols of two ascii table.}
\value{An ascii object.}
\author{David Hajage}
\arguments{\item{x}{an ascii table}
\item{y}{another ascii table}
\item{caption}{see \code{?ascii}}
\item{caption.level}{see \code{?ascii}}
\item{frame}{see \code{?ascii}}
\item{grid}{see \code{?ascii}}
\item{col.width}{see \code{?ascii}}
\item{width}{see \code{?ascii}}}
