% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.r
\name{cbind.ascii}
\alias{cbind.ascii}
\title{Cbind two ascii objects}
\usage{
\method{cbind}{ascii}(
  ...,
  caption = NULL,
  caption.level = NULL,
  frame = NULL,
  grid = NULL,
  col.width = 1,
  width = 0
)
}
\arguments{
\item{...}{ascii objects}

\item{caption}{see \code{?ascii}}

\item{caption.level}{see \code{?ascii}}

\item{frame}{see \code{?ascii}}

\item{grid}{see \code{?ascii}}

\item{col.width}{see \code{?ascii}}

\item{width}{see \code{?ascii}}
}
\value{
An \code{"asciiCbind"} object.
}
\description{
Cbind two ascii objects
}
\details{
This function binds cols of two ascii table.
}
\author{
David Hajage
}
