\name{convert}
\alias{convert}
\title{convert a file to several output using asciidoc...}
\usage{convert(input, destination, format="xhtml", encoding, latexmath=FALSE,
    asciimath=FALSE, cmd, cygwin=FALSE)}
\description{convert a file to several output using asciidoc}
\author{David Hajage}
\arguments{\item{input}{input file}
\item{destination}{output file (no extension)}
\item{format}{format of the output (chunked, epub, htmlhelp, pdf, text, xhtml, slidy, odt, dvi, ps, tex, docbook, asciidoc)}
\item{encoding}{encoding format of input file}
\item{latexmath}{use latexmath attribute}
\item{asciimath}{use asciimath attribute}
\item{cmd}{eventually define the asciidoc command}
\item{cygwin}{do you use asciidoc through cygwin}}
