\name{fig}
\alias{fig}
\title{Insert figure...}
\usage{fig(file, graph, format, ...)
}
\description{Insert figure}
\details{\code{graph} can be used with \code{export} function to insert an R graphic.}
\alias{graph}
\value{A fig object}
\author{David Hajage}
\arguments{\item{file}{character string (}
\item{graph}{a recordedplot, a lattice plot, a ggplot, or an expression producing a plot (optional if the file already exists)}
\item{format}{jpg, png or pdf (or guessed with the file name)}
\item{...}{additional arguments (passed to format options)}
}

