% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.asciidoc.r
\name{header.asciidoc}
\alias{header.asciidoc}
\title{header.asciidoc}
\usage{
header.asciidoc(
  caption = NULL,
  caption.level = NULL,
  frame = NULL,
  grid = NULL,
  col.width = 1,
  width = 0
)
}
\arguments{
\item{caption}{caption}

\item{caption.level}{caption.level}

\item{frame}{frame}

\item{grid}{grid}

\item{col.width}{col.width}

\item{width}{width}
}
\description{
header.asciidoc
}
\keyword{internal}
