% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.r
\name{interleave.data.frame}
\alias{interleave.data.frame}
\title{interleave.data.frame}
\usage{
interleave.data.frame(..., byrow = TRUE, pretty.rownames = TRUE)
}
\arguments{
\item{...}{...}

\item{byrow}{byrow}

\item{pretty.rownames}{pretty.rownames}
}
\description{
interleave.data.frame
}
\keyword{internal}
