\name{out}
\alias{out}
\title{Export R objects...}
\usage{out(x, results="verbatim")
}
\description{Export R objects}
\details{\code{out} can be used with \code{export} function to insert an R results}
\value{An out object}
\author{David Hajage}
\arguments{\item{x}{an R object}
\item{results}{if \code{'verbatim'}, the output is included in a verbatim environment. If \code{'latex'}, the output is taken to be already proper latex markup and included as is.}
}

