\name{paragraph}
\alias{paragraph}
\title{Create a paragraph...}
\usage{paragraph(..., new=TRUE)}
\description{Create a paragraph}
\details{\code{paragraph} can be used with \code{export} function to add\dots a paragraph}
\value{A paragraph object.}
\author{David Hajage}
\arguments{\item{...}{strings composing the paragraph}
\item{new}{whether to create a new paragraph or to continue a preceding one}}
