\name{paste.matrix}
\alias{paste.matrix}
\title{paste...}
\usage{paste.matrix(..., sep=" ", transpose.vector=FALSE, collapse=NULL, byrow=FALSE)
}
\description{paste.matrix}
\keyword{internal}
\arguments{\item{...}{...}
\item{sep}{sep}
\item{transpose.vector}{transpose.vector}
\item{collapse}{collapse}
\item{byrow}{byrow}
}

