\name{print.fig}
\alias{print.fig}
\title{Print an graph object...}
\usage{print.fig(x, backend=getOption("asciiBackend"), ...)
}
\description{Print an graph object}
\details{Print an graph object}
\author{David Hajage}
\arguments{\item{x}{an graph object}
\item{backend}{ascii backend}
\item{...}{not used}
}

