% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{print.out}
\alias{print.out}
\title{Print an out object}
\usage{
\method{print}{out}(x, backend = getOption("asciiBackend"), ...)
}
\arguments{
\item{x}{an out object}

\item{backend}{ascii backend}

\item{...}{not used}
}
\description{
Print an out object
}
\author{
David Hajage
}
