\name{print.section}
\alias{print.section}
\title{Print a section object}
\usage{
\method{print}{section}(x, backend = getOption("asciiBackend"), ...)
}
\arguments{
  \item{x}{a section object}

  \item{backend}{ascii backend}

  \item{...}{not used}
}
\description{
Print a section object
}
\author{
David Hajage
}

