\name{print.section}
\alias{print.section}
\title{Print a section object...}
\usage{print.section(x, backend=getOption("asciiBackend"), ...)
}
\description{Print a section object}
\details{Print a section object}
\author{David Hajage}
\arguments{\item{x}{a section object}
\item{backend}{ascii backend}
\item{...}{not used}
}

