% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{print.section}
\alias{print.section}
\title{Print a section object}
\usage{
\method{print}{section}(x, backend = getOption("asciiBackend"), ...)
}
\arguments{
\item{x}{a section object}

\item{backend}{ascii backend}

\item{...}{not used}
}
\description{
Print a section object
}
\author{
David Hajage
}
