\name{print.verbatim}
\alias{print.verbatim}
\title{Print a verbatim object...}
\usage{print.verbatim(x, backend=getOption("asciiBackend"), ...)
}
\description{Print a verbatim object}
\details{Print a verbatim object}
\author{David Hajage}
\arguments{\item{x}{a verbatim object}
\item{backend}{ascii backend}
\item{...}{not used}
}

