% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{print.verbatim}
\alias{print.verbatim}
\title{Print a verbatim object}
\usage{
\method{print}{verbatim}(x, backend = getOption("asciiBackend"), ...)
}
\arguments{
\item{x}{a verbatim object}

\item{backend}{ascii backend}

\item{...}{not used}
}
\description{
Print a verbatim object
}
\author{
David Hajage
}
