\name{replace}
\alias{replace}
\title{replace}
\usage{
replace(backend = getOption("asciiBackend"), cygwin = FALSE, i, f = NULL, d = NULL, e = NULL, O = NULL)
}
\arguments{
  \item{backend}{backend}

  \item{cygwin}{cygwin}

  \item{i}{i}

  \item{f}{f}

  \item{d}{d}

  \item{e}{e}

  \item{O}{O}
}
\description{
replace
}
\author{
David Hajage
}
\keyword{internal}

