\name{replace}
\alias{replace}
\title{replace...}
\usage{replace(backend=getOption("asciiBackend"), cygwin=FALSE, i, f, d, e, O)
}
\description{replace}
\keyword{internal}
\author{David Hajage}
\arguments{\item{backend}{backend}
\item{cygwin}{cygwin}
\item{i}{i}
\item{f}{f}
\item{d}{d}
\item{e}{e}
\item{O}{O}
}

