% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RweaveAscii.r
\name{RweaveAsciiSetup}
\alias{RweaveAsciiSetup}
\title{RweaveAsciiSetup}
\usage{
RweaveAsciiSetup(
  file,
  syntax,
  output = NULL,
  quiet = FALSE,
  debug = FALSE,
  extension = "txt",
  backend = "asciidoc",
  openSchunk = "----",
  closeSchunk = "\\n----\\n",
  openSinput = "",
  closeSinput = "",
  openSoutput = "\\n",
  closeSoutput = "",
  indent = "",
  openInclude = "include::",
  closeInclude = ".txt[]",
  openFig = "image::",
  closeFig = "[]",
  ...
)
}
\arguments{
\item{file}{file}

\item{syntax}{syntax}

\item{output}{output}

\item{quiet}{quite}

\item{debug}{debug}

\item{...}{...}

\item{stylepath}{stylepath}
}
\description{
RweaveAsciiSetup
}
\keyword{internal}
