% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RweaveAscii.r, R/weaverAscii.r
\name{RweavePandoc}
\alias{RweavePandoc}
\alias{RtanglePandoc}
\alias{RweavePandocOptions}
\alias{RweavePandocFinish}
\alias{RweavePandocWritedoc}
\alias{RweavePandocSetup}
\alias{RweavePandocRuncode}
\alias{weaverPandoc}
\alias{weaverPandocSetup}
\title{A driver to parse Pandoc noweb files with Sweave tool
This driver parses Pandoc files containing R code and replace pieces of code
with their output.}
\usage{
RweavePandoc()

RweavePandocSetup(
  file,
  syntax,
  output = NULL,
  quiet = FALSE,
  debug = FALSE,
  extension = "md",
  backend = "pandoc",
  openSchunk = "~~~~~~{.R}",
  closeSchunk = "\\n~~~~~~~~~~~\\n\\n",
  openSinput = "",
  closeSinput = "",
  openSoutput = "\\n",
  closeSoutput = "",
  indent = "",
  openInclude = "",
  closeInclude = "",
  openFig = "![](",
  closeFig = ")",
  ...
)

weaverPandocSetup(
  file,
  syntax,
  output = NULL,
  quiet = FALSE,
  debug = FALSE,
  extension = "md",
  backend = "pandoc",
  openSchunk = "~~~~~~{.R}",
  closeSchunk = "\\n~~~~~~~~~~~\\n\\n",
  openSinput = "",
  closeSinput = "",
  openSoutput = "\\n",
  closeSoutput = "",
  indent = "",
  openInclude = "",
  closeInclude = "",
  openFig = "![](",
  closeFig = ")",
  use.cache = TRUE,
  ...
)

weaverPandoc()
}
\arguments{
\item{file}{file}

\item{syntax}{syntax}

\item{output}{output}

\item{quiet}{quite}

\item{debug}{debug}

\item{...}{...}

\item{use.cache}{use.cache}

\item{stylepath}{stylepath}
}
\value{
None value is returned. From a .Rnw noweb file, the corresponding
  .md is produced (as eventuals files for graphs).
}
\description{
A driver to parse Pandoc noweb files with Sweave tool
This driver parses Pandoc files containing R code and replace pieces of code
with their output.

RweavePandocSetup

weaverPandocSetup

weaverPandoc
}
\note{
In order to work properly, noweb codes have to be located at the
  beginning of a line (no indentation).

Compare with RweaveLatex driver, RweavePandoc provides one new option :
  \code{format} to choose the format of figure that will be inserted in the
  final document.

In addition, \code{cache} option from \code{weaver}
  package is also available with \code{weaverPandoc} driver.

A wrapper for \code{Sweave} can be used, named \code{Pandoc}.
}
\examples{
  \dontrun{
Pandoc("file.Rnw")
  }

}
\seealso{
\code{\link[utils]{Sweave}}, \code{\link{Pandoc}}
}
\author{
David Hajage \email{dhajage@gmail.com} Matti Pastell \email{matti.pastell@helsinki.fi}
}
\keyword{IO}
\keyword{file}
\keyword{internal}
