\name{section}
\alias{section}
\title{Create a section...}
\usage{section(caption, caption.level=1)
}
\description{Create a section}
\details{\code{section} can be used with \code{export} function to add\dots a section}
\value{A section object.}
\author{David Hajage}
\arguments{\item{caption}{a string}
\item{caption.level}{caption level}
}

