% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{sexpr}
\alias{sexpr}
\title{Insert an inline R result}
\usage{
sexpr(x)
}
\arguments{
\item{x}{an R results (of length one)}
}
\value{
A sexpr object.
}
\description{
\code{sexpr} can be used with \code{export} function to insert an inline R results
}
\author{
David Hajage
}
