\name{show.asciidoc.list}
\alias{show.asciidoc.list}
\title{show...}
\usage{show.asciidoc.list(x, caption, caption.level, list.type="bullet", ...)}
\description{show.asciidoc.list}
\keyword{internal}
\arguments{\item{x}{x}
\item{caption}{caption}
\item{caption.level}{caption.level}
\item{list.type}{list.type}
\item{...}{...}}
