\name{show.org.list}
\alias{show.org.list}
\title{show...}
\usage{show.org.list(x, caption=NULL, caption.level=NULL, list.type="bullet", ...)
}
\description{show.org.list}
\keyword{internal}
\arguments{\item{x}{x}
\item{caption}{caption}
\item{caption.level}{caption.level}
\item{list.type}{list.type}
\item{...}{...}
}

