\name{show.rest.list}
\alias{show.rest.list}
\title{show...}
\usage{show.rest.list(x, caption, caption.level, list.type="bullet", ...)
}
\description{show.rest.list}
\keyword{internal}
\arguments{\item{x}{x}
\item{caption}{caption}
\item{caption.level}{caption.level}
\item{list.type}{list.type}
\item{...}{...}
}

