\name{verbatim}
\alias{verbatim}
\title{Create a verbatim paragraph...}
\usage{verbatim(...)
}
\description{Create a verbatim paragraph}
\details{\code{verbatim} can be used with \code{export} function to add a verbatim paragraph}
\value{A verbatim object.}
\author{David Hajage}
\arguments{\item{...}{strings composing the paragraph (line by line)}
}

